CantidadMaxima que se podrá decrementar

+++ src/main/java/es/iesjandula/reaktor/bookings_server/dto/RecursoCantMaxDto.java
@@ -1,5 +1,7 @@
 package es.iesjandula.reaktor.bookings_server.dto;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -12,6 +14,6 @@ public class RecursoCantMaxDto
 
 	private String recurso;
 
-	private Integer cantMax;
+	private BigDecimal cantMax;
 
 }
